# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EWAMA Properties Ltd API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated

  # ─── DASHBOARD ──────────────────────────────────────────
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List articles
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, published]
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Articles list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"

    post:
      operationId: createArticle
      tags: [articles]
      summary: Create an article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get an article
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"

    patch:
      operationId: updateArticle
      tags: [articles]
      summary: Update an article
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"

    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete an article
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /articles/{id}/publish:
    post:
      operationId: publishArticle
      tags: [articles]
      summary: Publish an article
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"

  # ─── ENQUIRIES ───────────────────────────────────────────
  /enquiries:
    get:
      operationId: listEnquiries
      tags: [enquiries]
      summary: List customer enquiries
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [unread, read, contacted, closed]
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Enquiries list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnquiryListResponse"

    post:
      operationId: createEnquiry
      tags: [enquiries]
      summary: Submit an enquiry (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnquiryInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
  /enquiries/{id}:
    get:
      operationId: getEnquiry
      tags: [enquiries]
      summary: Get an enquiry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"

    patch:
      operationId: updateEnquiry
      tags: [enquiries]
      summary: Update enquiry status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnquiryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
...
  /media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List media files
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 24
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaListResponse"

    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload a media file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaInput"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaFile"
  /media/{id}:
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete a media file
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SETTINGS ────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string

        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
        token:
          type: string
    AdminUser:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ─── DASHBOARD ─────────────────────────────────────────
...
    Article:
      type: object
      required: [id, title, slug, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        featuredImage:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        author:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published]
        readingTime:
          type: ["integer", "null"]
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ArticleInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 2
        slug:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        featuredImage:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        author:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published]
          default: draft
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
    ArticleUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        featuredImage:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        author:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, published]
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
    ArticleListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── ENQUIRY ───────────────────────────────────────────
    Enquiry:
      type: object
      required: [id, name, email, phone, message, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        propertyId:
          type: ["integer", "null"]
        propertyName:
          type: ["string", "null"]
        message:
          type: string
        status:
          type: string
          enum: [unread, read, contacted, closed]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    EnquiryInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
        email:
          type: string

        phone:
          type: ["string", "null"]
        propertyId:
          type: ["integer", "null"]
        message:
          type: string
    EnquiryUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [unread, read, contacted, closed]
        notes:
          type: ["string", "null"]
    EnquiryListResponse:
      type: object
      required: [data, total, page, limit, unreadCount]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Enquiry"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        unreadCount:
          type: integer

    # ─── SITE VISIT ────────────────────────────────────────
...
    MediaFile:
      type: object
      required: [id, fileName, url, mimeType, size, createdAt]
      properties:
        id:
          type: integer
        fileName:
          type: string
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        mimeType:
          type: string
        size:
          type: integer
        altText:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    MediaInput:
      type: object
      required: [fileName, url, mimeType, size]
      properties:
        fileName:
          type: string
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        mimeType:
          type: string
        size:
          type: integer
        altText:
          type: ["string", "null"]
    MediaListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MediaFile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── SETTINGS ──────────────────────────────────────────
...