# lib/api-zod/src/generated/types/mediaListResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * EWAMA Properties Ltd API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaFile } from './mediaFile';

export interface MediaListResponse {
  data: MediaFile[];
  total: number;
  page: number;
  limit: number;
}